# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API that queues Whisper-style audio transcription jobs via Redis. Versioned routes live under `/api/v1/...`.",
        "title": "Whisper Fiber API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/transcribe": {
            "post": {
                "description": "Accepts an audio file in the multipart form field `audio`, generates a `task_id` (UUID), and enqueues the job. Use WAV/MP3 or other formats supported by your downstream pipeline.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audio"
                ],
                "summary": "Speech-to-text enqueue",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Audio file (e.g. wav/mp3)",
                        "name": "audio",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON body includes task_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing audio field or invalid multipart form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to read the uploaded file or enqueue the job",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/echo": {
            "get": {
                "description": "Returns **`text/plain`**. If the request has a body, it is echoed as-is. If the body is empty, the query **`q`** is echoed when present; otherwise the response body is empty (HTTP 200, no error).",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "debug"
                ],
                "summary": "Echo string",
                "parameters": [
                    {
                        "type": "string",
                        "description": "String to echo when the request body is empty (Swagger: use GET /echo)",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request body bytes, or q, or empty",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns an aggregated dependency status (e.g. `redis_status`) and optional Triton/Whisper backend fields (`whisper_triton_status`, `whisper_triton_message`) when WHISPER_TRITON_HTTP_URL is set. Responds with HTTP 503 when Redis is unavailable.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "All checks passed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service temporarily unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}